+++ example/src/components/BackgroundPlaybackToggle.tsx
@@ -1,15 +1,16 @@
 import { GlyphToy } from '@missingcore/music-glyph-toys';
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { StyleSheet, Switch, Text, View } from 'react-native';
+import AudioBrowser from 'react-native-audio-browser';
 
-import { setPlayerOptions } from '../utils/react-native-track-player';
+import { setPlayerOptions } from '../utils/react-native-audio-browser';
 
 export function BackgroundPlaybackToggle() {
   const [isEnabled, setIsEnabled] = useState(true);
   const killAppOnDismiss = useRef(false);
 
   const onChange = useCallback((value: boolean) => {
-    setPlayerOptions(value);
+    AudioBrowser.updateOptions(setPlayerOptions(value));
     setIsEnabled(value);
     killAppOnDismiss.current = !value;
   }, []);