+++ example/src/utils/react-native-audio-browser.ts
@@ -1,0 +1,18 @@
+export function setPlayerOptions(continuePlaybackOnDismiss = true) {
+  return {
+    android: {
+      appKilledPlaybackBehavior: continuePlaybackOnDismiss
+        ? 'continue-playback'
+        : 'stop-playback-and-remove-notification',
+    },
+    capabilities: {
+      stop: false,
+      jumpForward: false,
+      jumpBackward: false,
+      favorite: false,
+      shuffleMode: false,
+      repeatMode: false,
+      playbackRate: false,
+    },
+  } as const;
+}